Breaking: Move util functions, automatically npm link when needed
