+++ all-configs.eslintrc.js
@@ -1,3 +1,5 @@
+require("./lib/util/link");
+
 // eslint-disable-next-line node/no-extraneous-require
 const {configs} = require("eslint-plugin-evelyn");
