+++ changes/record-changes.eslintrc.js
@@ -1,6 +1,7 @@
 const lodash = require("lodash");
+require("../lib/util/link");
 // eslint-disable-next-line node/no-extraneous-require
-require("eslint-plugin-evelyn/lib/save-config")(
+require("eslint-plugin-evelyn/lib/util/save-config")(
 	"./changes/record-changes.json",
 	data => lodash.cloneDeepWith(
 		JSON.parse(JSON.stringify(data)),
